set up base for overview subpage
