+++ src/app/Features/overview/pages/home/home.component.ts
@@ -42,16 +42,7 @@ import { OverviewView } from '@features/overview/views/overview-view/overview-vi
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.css'] 
 })
-export class HomeComponent implements OnInit {
-    years: number[] = [2023, 2024, 2025];
-    selectedYear = new Date().getFullYear(); 
-
-    chartModes: string[] = ['default', 'by-category'];
-    selectedChartMode = 'default';
-
-    monthlySummary$!: Observable<any>;
-    monthlyChartSummary$!: Observable<any>;
-
+export class HomeComponent {
     tabs = [
         { label: 'Overview', component: OverviewView },
         { label: 'Expenses', component: ExpensesView },
@@ -60,85 +51,6 @@ export class HomeComponent implements OnInit {
 
     selectedIndex = 0;
   
-    constructor(private financialService: FinancialService) {}
-
-    ngOnInit(): void {
-        this.loadData(this.selectedYear, this.selectedChartMode);
-    }
-
-    onYearChange(year: number): void {
-        this.selectedYear = year;
-        this.loadData(year, this.selectedChartMode);
-    }
-
-    onChartModeChange(mode: string): void {
-        this.selectedChartMode = mode;
-        this.loadData(this.selectedYear, mode);
-    }
-
-    private loadData(year: number, mode: string): void {
-        this.monthlySummary$ = this.financialService.getMonthlyIncomeAndOutcome(year);
-
-        if (mode === 'default') {
-            this.monthlyChartSummary$ = this.buildChartSummary(this.monthlySummary$);
-        }
-
-        if (mode === 'by-category') {
-            this.monthlyChartSummary$ = this.buildChartSummary(this.monthlySummary$);
-        }
-
-    }
-
-
-    /**
-     * Transforms a stream of `DefaultSummary[]` into a chart-friendly structure
-     * for use with ngx-charts.
-     *
-     * Each `DefaultSummary` item is mapped into two series:
-     *  - **Income**: values from the `income` property
-     *  - **Expenses**: values from the `expense` property
-     *
-     * @param source$ Observable emitting arrays of `DefaultSummary` objects.
-     * @returns Observable emitting a chart data array, where each entry has:
-     *   - `name`: the label of the series ("Income" or "Expenses")
-     *   - `series`: an array of `{ name: string, value: number }` points
-     */
-    private buildChartSummary(source$: Observable<DefaultSummary[]>): Observable<any[]> {
-        return source$.pipe(
-            map(data => [
-                this.buildSeries('Income', data, item => item.income),
-                this.buildSeries('Expenses', data, item => item.expense)
-            ])
-        );
-    }
-
-    /**
-     * Builds a single chart series object for use with ngx-charts.
-     *
-     * Iterates over an array of `DefaultSummary` objects and transforms
-     * each entry into a `{ name, value }` point, where:
-     *  - `name` is the month label from the `DefaultSummary`
-     *  - `value` is derived by applying the provided `selector` function
-     *
-     * @param name - The label of the chart series (e.g. "Income", "Expenses").
-     * @param data - The list of `DefaultSummary` items to transform.
-     * @param selector - A function that extracts the numeric value from
-     *   each `DefaultSummary` (e.g. `item => item.income`).
-     *
-     * @returns An object with:
-     *   - `name`: the series label
-     *   - `series`: an array of `{ name: string, value: number }` points
-     */
-    private buildSeries(name: string, data: DefaultSummary[], selector: (item: DefaultSummary) => number) {
-        return {
-            name,
-            series: data.map(item => ({
-                name: item.month,
-                value: selector(item)
-            }))
-        };
-    }
-
     private analytics = inject(AnalyticsService);
 
     // state (simpele variant)
